Move resolveModule file into modules dir
